# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tenants": {
      "get": {
        "operationId": "TenantsController_list[0]_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a tenant (Platform Administrator)",
        "tags": [
          "tenants"
        ]
      }
    },
    "/tenants": {
      "get": {
        "operationId": "TenantsController_list[1]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a tenant (Platform Administrator)",
        "tags": [
          "tenants"
        ]
      }
    },
    "/v1/tenants/{tenantId}": {
      "patch": {
        "operationId": "TenantsController_update[0]_v1",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tenants"
        ]
      }
    },
    "/tenants/{tenantId}": {
      "patch": {
        "operationId": "TenantsController_update[1]",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tenants"
        ]
      }
    },
...
    "/v1/modules": {
      "get": {
        "operationId": "ModulesController_list[0]_v1",
        "parameters": [
          {
            "name": "application_code",
            "required": false,
            "in": "query",
            "schema": {
              "example": "rms",
              "type": "string"
            }
          },
          {
            "name": "is_default",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modules with their permissions (14 defaults + application modules)",
        "tags": [
          "modules"
        ]
      },
      "post": {
        "operationId": "ModulesController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a custom module (is_default = false) and its permissions",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules": {
      "get": {
        "operationId": "ModulesController_list[1]",
        "parameters": [
          {
            "name": "application_code",
            "required": false,
            "in": "query",
            "schema": {
              "example": "rms",
              "type": "string"
            }
          },
          {
            "name": "is_default",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modules with their permissions (14 defaults + application modules)",
        "tags": [
          "modules"
        ]
      },
      "post": {
        "operationId": "ModulesController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a custom module (is_default = false) and its permissions",
        "tags": [
          "modules"
        ]
      }
    },
    "/v1/permissions": {
      "get": {
        "operationId": "PermissionsController_list[0]_v1",
        "parameters": [
          {
            "name": "module_code",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ROLE_MGMT",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permissions"
        ]
      },
      "post": {
        "operationId": "PermissionsController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an action to a module; permission_code = <MODULE_CODE>_<ACTION>",
        "tags": [
          "permissions"
        ]
      }
    },
    "/permissions": {
      "get": {
        "operationId": "PermissionsController_list[1]",
        "parameters": [
          {
            "name": "module_code",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ROLE_MGMT",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permissions"
        ]
      },
      "post": {
        "operationId": "PermissionsController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an action to a module; permission_code = <MODULE_CODE>_<ACTION>",
        "tags": [
          "permissions"
        ]
      }
    },
    "/v1/roles": {
      "get": {
        "operationId": "RolesController_list[0]_v1",
        "parameters": [
          {
            "name": "scope_level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CORE",
                "BUSINESS_UNIT",
                "UTILITY"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "RolesController_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a custom role at or below your workspace level (e.g. \"Zone Manager\")",
        "tags": [
          "roles"
        ]
      }
    },
    "/roles": {
      "get": {
        "operationId": "RolesController_list[1]",
        "parameters": [
          {
            "name": "scope_level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CORE",
                "BUSINESS_UNIT",
                "UTILITY"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "RolesController_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a custom role at or below your workspace level (e.g. \"Zone Manager\")",
        "tags": [
          "roles"
        ]
      }
    },
    "/v1/roles/{roleId}": {
      "get": {
        "operationId": "RolesController_get[0]_v1",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_rename[0]_v1",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roles"
        ]
      }
    },
    "/roles/{roleId}": {
      "get": {
        "operationId": "RolesController_get[1]",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_rename[1]",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "roles"
        ]
      }
    },
    "/v1/role-permissions": {
      "post": {
        "operationId": "RolePermissionsController_change[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grant or revoke permissions on a role (no escalation beyond your own permissions)",
        "tags": [
          "roles"
        ]
      }
    },
    "/role-permissions": {
      "post": {
        "operationId": "RolePermissionsController_change[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grant or revoke permissions on a role (no escalation beyond your own permissions)",
        "tags": [
          "roles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Miqaat Core Identity Authorization Service",
    "description": "Dynamic Business Unit / Utility / Application / Environment / Module / Role / Permission model, client registry and server-side authorization checks. No API keys: every call carries an RS256 bearer JWT verified via JWKS - service tokens (typ client-authentication+jwt) from registered service principals, or administrator access tokens (typ at+jwt) issued by Identity Federation. Never stores credentials.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256, verified via JWKS)",
        "type": "http"
      }
    },
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Miqaat"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "default": "active"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "CreateModuleDto": {
        "type": "object",
        "properties": {
          "module_code": {
            "type": "string",
            "example": "RMS_REGISTRATION"
          },
          "module_name": {
            "type": "string",
            "example": "RMS Registration"
          },
          "application_code": {
            "type": "string",
            "example": "rms",
            "description": "Owning application; omit for a Core Portal module"
          },
          "actions": {
            "type": "array",
            "example": [
              "view",
              "create",
              "edit"
            ],
            "items": {
              "type": "string",
              "enum": [
                "view",
                "create",
                "edit",
                "delete",
                "approve",
                "export"
              ]
            }
          }
        },
        "required": [
          "module_code",
          "module_name"
        ]
      },
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "module_code": {
            "type": "string",
            "example": "ROLE_MGMT"
          },
          "action": {
            "type": "string",
            "enum": [
              "view",
              "create",
              "edit",
              "delete",
              "approve",
              "export"
            ],
            "example": "export"
          }
        },
        "required": [
          "module_code",
          "action"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "role_name": {
            "type": "string",
            "example": "Zone Manager"
          },
          "scope_level": {
            "type": "string",
            "enum": [
              "CORE",
              "BUSINESS_UNIT",
              "UTILITY"
            ],
            "example": "UTILITY"
          },
          "permission_codes": {
            "example": [
              "DASHBOARD_VIEW",
              "TICKET_MGMT_VIEW",
              "TICKET_MGMT_CREATE"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tenant_id": {
            "type": "string",
            "description": "Only for ADMIN service principals"
          }
        },
        "required": [
          "role_name",
          "scope_level"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "role_name": {
            "type": "string",
            "example": "Zone Manager (North)"
          }
        },
        "required": [
          "role_name"
        ]
      },
      "RolePermissionsDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "string"
          },
          "permission_codes": {
            "example": [
              "TICKET_MGMT_EDIT"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "action": {
            "type": "string",
            "enum": [
              "GRANT",
              "REVOKE"
            ],
            "default": "GRANT"
          }
        },
        "required": [
          "role_id",
          "permission_codes"
        ]
      },
...
    }
  }
}